+++ CHANGELOG.md
@@ -56,6 +56,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    * Changed `fname` from a kwarg to an arg in `pysat.Instruments.to_netcdf4`
    * Deprecated support for 2D pandas datasets
    * Deprecated `pysat.instruments.pysat_testing_xarray`
+   * Deprecated usage of higher order metadata
 * Documentation
    * Moved logo to 'docs\images'
    * Improved consistency of headers throughout documentation